requirements.txt
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
h11==0.14.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
